🐛 fix some bug
