+++ apps/web/components/page-components/editor/Middle.tsx
@@ -6,14 +6,14 @@ import ChangePositionComponent from "../../base/ChangePositionComponent";
 import ResizeComponent from "../../base/ResizeComponent";
 
 function Middle(props: any) {
-  const { Elements, setIsElement, setMode, mode } = UseElementStore();
+  const { Elements, setIsElement, currentElement,setCurrentElement } = UseElementStore();
 
   return (
     <button
       className="bg-[#f0f2f5] w-3/5 flex justify-center items-center flex-col"
       onClick={() => {
         setIsElement(false);
-        setMode(true);
+        setCurrentElement('');
       }}
     >
       <h3>海报区域</h3>
@@ -23,22 +23,21 @@ function Middle(props: any) {
         style={{ width: "375px", height: "667px", position: "relative", overflow: "auto" }}
       >
         {Elements.map((item: any) =>
-          mode ? (
-            <div
-              key={item.id}
-              className={` ${item.isHidden ? "invisible" : ""}`}
-            >
-              <ChangePositionComponent item={item} />
-            </div>
-          ) : (
+          item.id == currentElement ? (
             <div
               key={item.id}
               className={` ${item.isHidden ? "invisible" : ""}`}
             >
               <ResizeComponent item={item} />
             </div>
+          ) : (
+            <div
+              key={item.id}
+              className={` ${item.isHidden ? "invisible" : ""}`}
+            >
+              <ChangePositionComponent item={item} />
+            </div>
           ),
-          
         )}
       </div>
     </button>
