+++ apps/web/package.json
@@ -19,7 +19,6 @@
         "@radix-ui/react-slot": "1.1.0",
         "@radix-ui/react-toggle": "^1.1.0",
         "antd": "5.20.0",
-        "array-move": "^4.0.0",
         "class-variance-authority": "0.7.0",
         "clsx": "2.1.1",
         "lucide-react": "0.424.0",
@@ -30,7 +29,6 @@
         "react-color": "^2.19.3",
         "react-dom": "18.3.1",
         "react-hook-form": "7.52.2",
-        "react-sortable-hoc": "^2.0.0",
         "socket.io-client": "4.7.5",
         "tailwind-merge": "2.4.0",
         "tailwindcss-animate": "1.0.7",
@@ -44,7 +42,6 @@
         "@types/react": "18.3.3",
         "@types/react-color": "^3.0.12",
         "@types/react-dom": "18.3.0",
-        "@types/react-sortable-hoc": "^0.7.1",
         "@types/three": "0.167.1",
         "@types/uuid": "10.0.0",
         "css-types": "1.1.0",
