+++ apps/web/utils/arrayMove.ts
@@ -1,0 +1,6 @@
+export function arrayMove(list: any[], fromIndex: number, toIndex: number) {
+  const element = list[fromIndex];
+  list.splice(fromIndex, 1);
+  list.splice(toIndex, 0, element);
+  return list;
+}
