+++ apps/web/utils/getParentElement.ts
@@ -1,0 +1,10 @@
+export function getParentElement(element: HTMLElement, className: string) {
+  while (element) {
+    if (element.classList && element.classList.contains(className)) {
+      return element;
+    } else {
+      element = element.parentNode as HTMLElement;
+    }
+  }
+  return null;
+}
